Add a new 'sync' command.
Usage: drush [options] sync <destination> <source>\n\nRsync the entire drupal directory (or subdirectory, using $source parameter) to <destination> using ssh. Excludes .svn directories. Useful for pushing copies of your tree to a staging server, for example. The most common case for specifying $source is for pointing to a 'files' directory when thats all you want to sync. <source> may be relative to Drupal root or an absolute path. If <source> is omitted, the whole Drupal tree is pushed.