1129332 - do not override builtin error_reporting level
instead of recording the "previous error_reporting() level" (which is
actually 0 when errors are on, weird), we record the "default
error_reporting() level" - either the system-wide defined in php.ini or
the one set before drush (for example by your script in
hook_drush_init() if you want to override it).

This way, you can have code in hook_drush_init() that just calls
error_reporting() to change the default and it will "just work".