671906 - show the password correctly in sql-connect
this fixes a regression introduced in the previous commit where the
secure credentials passing was also applied to sql-connect. this
obviously breaks because the tempfile is gone.

this also sets sql-sync to pass the password around insecurely if it
is dealing with remote hosts

this is kind of dirty, and would be better accomplished by turning
sql_get_credentials() into a parser that would return an array that
could be *then* fixed by calling functions, as necessary.