Fix sql-sync not propagating options to sub-commands
Currently, the following example at examples/example.aliases.drushrc.php
does not work:

```
 #$aliases['live'] = array(
 #    'parent' => '@server,@dev',
 #    'uri' => 'mydrupalsite.com',
 #     'target-command-specific' => array (
 #       'sql-sync' => array (
 #         'skip-tables-list' => 'comments',
 #       ),
 #     ),
 #  );
```

The reason is that skip-tables-list is an option for sql-dump, not
for sql-sync. The above would only work if the setting was set to
the sql-dump command instead of the sql-sync command. However, this
may be desired just for sql-sync so it is not a good solution.
Besides, this must be frustrating developers which do not
understand why tables are not ignored.

We should propagate sql-sync options to its sub-commands
such as sql-dump or sql-sanitize. Perhaps on a follow up we could
remove options which belong to sub-commads defined at sql-sync
and instead use allow-additional-options => ('sql-create', 'sql-dump',
'sql-sanitize').