Replicate ini tests for YAML format in make tests array.
Most elements in the makefile tests array should always be identical (e.g.
'options', 'md5', etc.) Those that should differ basically just append a
string to indicate that it's a YAML-formatted makefile. So, we save a bunch
of duplication, that'd be hard to keep in sync, by simply duplicating the
tests programmatically.