Fixes drush_get_runner() when @self is a param
Consider the following command:

    dev $ drush config-pull @prod @self

Since this is being run on the dev server, I would expect this to mean "Copy config from prod to dev." Unfortunately, this can mean "copy config from prod to prod". To understand why, let's examine the these two lines which run the rsync operations:

    $runner = drush_get_runner($source, $destination);
    drush_invoke_process($runner, 'core-rsync', array("$source:$export_path", "$destination:%config-$label"));

In this code, $source is '@prod' and $destination is '@self'. The gotcha is that the evaluation of @self happens at $runner, which can be either @prod or @dev. If $runner is @prod, then we have a @prod => @prod transfer. Not good. Now fixed.