Remove usage of sudo from install alt docs
The steps for installing Drush for all users via composer do not require
elevated privileges unless writing to a privileged directory path.
Afterwards nobody needs to prefix the Drush command with sudo to use
Drush. We're removing references to sudo here so that this is clear. You
dont need to run Drush with sudo for this to work.