Fix PHP pcntl_exec() Error in install instructions
The current install instructions generate a PHP pcntl_exec() permissions
error when run with PHP 5.6. Some people are incorrectly assuming this
means drush needs to be installed as root.

Executing the php call to `drush core-status` from outside the path
where the drush PHAR file is located fixes the issue.