Fix the sitelist when the site is symlinked
- "$real_root = drush_locate_root($filename)"
always return true even if the site directory is symlinked to a directory outside drupal root. This fixes cases where some sites are symlinked. For example: If you have a site that is symlinked `vanilla -> ../../vanilla/src` where the target is a directory outside the drupal root, the function drush_locate_root return true we end up with a wrong alias.