Stop all processing of queue on SuspendQueueException
If a SuspendQueueException is thrown when processing an item the current
implementation will release the item and log an error. Processing of the
queue continues.

According to the documentation of the SuspendQueueException throwing 
this exception indicates that processing of the whole queue should be skipped.

This change breaks the queue item processing loop entirely after logging
the error.

The associated code comment is also updated as it does not make sense.
The queue run command only runs for a single queue so there is nothing
to skip to.