Highlight need to run composer in D7 manual install docs
You still need to run composer install at the very end of a Drush 7 manual install, to populate the vendor directory. 
This is a bit confusing if you're scanning the install methods quickly and have chosen not to use the "composer" methods.  
Also, adds exact error message if not run, to aid those searching.