Add an 'ssh.tty' option to site aliased. (#3648)
* Add an 'ssh.tty' option to site aliased.

Usually Drush decides whether or not to create a tty for remote ssh commands based on whether the current Drush command is running in interactive mode or not. When in interactive mode, a tty is created; otherwise, no tty is created.

Sometimes, a remote server will not grant the ssh user the permission to create a tty. In this instance, it is possible to force Drush to never create a tty via this new option.

* Allow ssh.tty to override interactive mode regardless of whether interactive is true or false.