Convert Drush Make to Composer more accurately (#4082)
This adds every library as a 'package' type to the repository array of Composer. 
It also removes the caret (^) in front of packages, so running `composer install` downloads the same versions as `drush make` does.
Furthermore, it uses the original patch name. Lastly, it also marks `composer/core` to be fully incompatible as opposed to only version 8.x.