Prevent patch context being ignored (#4193)
The initial 'git apply' command does not employ the -C option to allow
fuzzy context matching, so neither should the fallback 'patch' command
be allowed to use its default fuzz factor.  Both commands should treat
context in a consistent manner.

By default 'patch' will allow a hunk to apply with only a single line
of the provided context matching either side, and by default drush make
will not show any of the output which indicates that such fuzzy matching
has taken place.

This makes it far too easy for inapplicable patches to be 'successfully'
(but incorrectly) applied, with no warning to the user.