Do not pass --root in sql-sync if it is not set in the alias file.
Drush sql-sync will add a "--root" option to the backend invoke call that fetches the database information. If the site alias record does not define a "root" path, then Drush will use the path to the root of the locally-bootstrapped Drupal site. This will not produce the correct results, because the local root path is usually meaningless on the remote server.

This commit fixes this bug. This bug has already been fixed in Drush 9 and Drush 10.