Fix issue #5534
Although it was suggested enclosing the password in backticks would resolve this.  Further testing indicated that it didn't work when the password contains a backtick.  Also, the quoting mentioned in my previous attempt at a PR regarding the database name doesn't really have anything to do with this in my opinion.  I could put a check in to see if that was set but again in my opinion there is never a reason to not perform what the fix is doing since it is always a possiblity that a user will use a backtick in their password.